Fix e2e tests
