--- e2e/test-api-client-esm/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "test-api-client-esm",
-  "version": "0.0.0",
-  "private": true,
-  "engines": {
-    "node": ">=18"
-  },
-  "type": "module",
-  "scripts": {
-    "start": "node ./index.js"
-  }
-}
