+++ test/e2e/plain-script.ts
@@ -1,5 +1,6 @@
 import { AutoblocksTracer } from '../../src';
 import { BaseEventEvaluator } from '../../src/testing';
+import { E2E_TESTS_MESSAGE_NAME } from './util';
 
 class SlowEvaluator1 extends BaseEventEvaluator {
   id = 'e2e-slow-evaluator-1';
@@ -23,7 +24,7 @@ class SlowEvaluator2 extends BaseEventEvaluator {
   const traceId = process.argv[2];
   const tracer = new AutoblocksTracer();
   console.log('Sending event at ', new Date().toISOString());
-  tracer.sendEvent('e2e-test-event-plain-script', {
+  tracer.sendEvent(E2E_TESTS_MESSAGE_NAME, {
     traceId,
     evaluators: [new SlowEvaluator1(), new SlowEvaluator2()],
   });
