+++ test/e2e/util.ts
@@ -1,0 +1,1 @@
+export const E2E_TESTS_MESSAGE_NAME = 'sdk.e2e';
